pyproject.toml
[tool.poetry]
name = "esch"
version = "0.1.28"
description = "A tool for converting markdown files with frontmatter"
authors = ["Noah Syrkis <noah@syrkis.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
markdown-analysis = "^0.0.5"
python-frontmatter = "^1.1.0"

[tool.poetry.scripts]
esch = "esch.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"